backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0
ftfy==6.1.1
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
tqdm==4.66.1

# Install CLIP separately: pip install git+https://github.com/openai/CLIP.git
